Changed UnitPrice data type from double to bigdecimal

+++ backend/src/main/java/io/github/Lordrap2002/inventory/api/dto/ProductRequestDTO.java
@@ -3,6 +3,7 @@ package io.github.Lordrap2002.inventory.api.dto;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public class ProductRequestDTO {
 
@@ -11,7 +12,7 @@ public class ProductRequestDTO {
 
     @NotNull(message = "The price is required")
     @Min(value = 0, message = "The price must be greater than or equal to 0")
-    private Double unitPrice;
+    private BigDecimal unitPrice;
 
     @NotNull(message = "The stock is required")
     @Min(value = 0, message = "The stock must be greater than or equal to 0")
@@ -27,11 +28,11 @@ public class ProductRequestDTO {
         this.name = name;
     }
 
-    public Double getUnitPrice() {
+    public BigDecimal getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Double unitPrice) {
+    public void setUnitPrice(BigDecimal unitPrice) {
         this.unitPrice = unitPrice;
     }
